backend/requirements.txt
fastapi
slowapi
uvicorn[standard]
python-dotenv
redis
cryptography
python-multipart